requirements.txt
flask==2.3.3
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.20
passlib==1.7.4
werkzeug==2.3.7
requests==2.31.0
ipwhois==1.2.0
dnspython==2.0.0
python-whois==0.8.0
Pillow==10.0.0
pytesseract==0.3.10
google-generativeai
openai
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-dotenv==1.0.0
PyMuPDF==1.23.5
PyPDF2==3.0.1
python-docx==0.8.11
phonenumbers==8.13.22
tldextract==5.1.2
deep-translator==1.11.4
